requirements.txt
python-docx==0.8.11
lxml==4.9.3
click==8.1.7